requirements.txt
--index-url http://pypi.douban.com/simple

Django
mysql-python
django-parsley
python-memcached
django_select2
django-tables2
django-tables3
django-mysqlpool
django-oauth-toolkit
uwsgi
pillow
requests
redis

django-tastypie
django-render-json
django-render-csv
